added validations to user sign up

+++ db/migrate/20180601180632_create_pairings.rb
@@ -2,7 +2,7 @@ class CreatePairings < ActiveRecord::Migration[5.2]
   def change
     create_table :pairings do |t|
       t.integer :comment_count
-      t.integer :user_rating
+      t.float :user_rating
     	t.integer :wine_id
     	t.integer :cookie_id
     	t.integer :user_id
